pyproject.toml
[project]
name = "prompt-engineering-playground"
version = "0.1.0"
description = ""
authors = [
    {name = "Aymane Cheikh",email = "aymanecheikh2002@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai (>=2.15.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "prompt_engineering_playground", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
